requirements.txt
langchain
ipykernel
pandas
numpy
langchain-openai
python-dotenv
langchain_community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
faiss-cpu
sentence_transformers
langchain_huggingface
langchain_chroma
duckdb
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
streamlit